+++ src/Components/Cockpit/Cockpit.module.css
@@ -1,0 +1,32 @@
+.red{
+  color: red;
+}
+
+.bold{
+  font-weight: bold;
+}
+
+.Cockpit button{
+  background-color:green;
+	font: inherit;
+	border: 1px solid red;
+	padding: 8px;
+  cursor: pointer;
+  color: white;
+}
+
+
+.Cockpit button:hover {
+  background-color: yellowgreen;
+  color: black;
+}
+
+.Cockpit button.Red{
+  background-color: red;
+  border: 1px solid green;
+}
+
+.Cockpit button.Red:hover{
+  background-color: salmon;
+  /* border: 1px solid green; */
+}
