+++ src/containers/App.module.css
@@ -1,36 +1,3 @@
 .App {
   text-align: center;
 }
-
-.red{
-  color: red;
-}
-
-.bold{
-  font-weight: bold;
-}
-
-.App button{
-  background-color:green;
-	font: inherit;
-	border: 1px solid red;
-	padding: 8px;
-  cursor: pointer;
-  color: white;
-}
-
-
-.App button:hover {
-  background-color: yellowgreen;
-  color: black;
-}
-
-.App button.Red{
-  background-color: red;
-  border: 1px solid green;
-}
-
-.App button.Red:hover{
-  background-color: salmon;
-  /* border: 1px solid green; */
-}
